Move protocol views load to gui module getProtocolViews, _loadProtocols, getCurrentProtocolView moved from Project (model) to viewprotocols
