+++ pyworkflow/gui/project/project.py
@@ -167,7 +167,6 @@ class ProjectWindow(ProjectBaseWindow):
             self.settings = proj.createSettings()
 
         self.generalCfg = self.settings.getConfig()
-        self.protCfg = proj.getCurrentProtocolView()
 
         return proj
