+++ pyworkflow/project/__init__.py
@@ -28,5 +28,4 @@
 from .project import Project, MissingProjectDbException
 from .manager import Manager, ProjectInfo
 from .config import (ProjectSettings, ProjectConfig, MenuConfig,
-                     ProtocolConfig, NodeConfig, NodeConfigList,
-                     Label, LabelsList, ProtocolTreeConfig)
+                     NodeConfig, NodeConfigList, Label, LabelsList)
