+++ pyworkflow/project/config.py
@@ -226,26 +226,6 @@ class MenuConfig(object):
         return len(self.childs) == 0
 
 
-class ProtocolConfig(MenuConfig):
-    """Store protocols configuration """
-
-    def __init__(self, text=None, value=None, **args):
-        MenuConfig.__init__(self, text, value, **args)
-        if 'openItem' not in args:
-            self.openItem.set(self.tag.get() != 'protocol_base')
-
-    def addSubMenu(self, text, value=None, shortCut=None, **args):
-        if 'icon' not in args:
-            tag = args.get('tag', None)
-            if tag == 'protocol':
-                args['icon'] = 'python_file.gif'
-            elif tag == 'protocol_base':
-                args['icon'] = 'class_obj.gif'
-
-        args['shortCut'] = shortCut
-        return MenuConfig.addSubMenu(self, text, value, **args)
-
-
 class NodeConfig(pwobj.Scalar):
     """ Store Graph node information such as x, y. """
 
@@ -417,224 +397,3 @@ class LabelsList(pwobj.List):
     def clear(self):
         pwobj.List.clear(self)
         self._labelDict.clear()
-
-
-class ProtocolTreeConfig:
-    """ Handler class that groups functions and constants
-    related to the protocols tree configuration.
-    """
-    ALL_PROTOCOLS = "All"
-    TAG_PROTOCOL_DISABLED = 'protocol-disabled'
-    TAG_PROTOCOL = 'protocol'
-    TAG_SECTION = 'section'
-    TAG_PROTOCOL_GROUP = 'protocol_group'
-    PLUGIN_CONFIG_PROTOCOLS = 'protocols.conf'
-
-    @classmethod
-    def getProtocolTag(cls, isInstalled):
-        """ Return the proper tag depending if the protocol is installed or not.
-        """
-        return cls.TAG_PROTOCOL if isInstalled else cls.TAG_PROTOCOL_DISABLED
-
-    @classmethod
-    def isAFinalProtocol(cls, v, k):
-        if (issubclass(v, pw.viewer.ProtocolViewer) or
-                v.isBase() or v.isDisabled()):
-            return False
-
-        # To remove duplicated protocol, ProtMovieAlignment turns into OF:
-        # ProtMovieAlignment = XmippProtOFAlignment
-        return v.__name__ == k
-
-    @classmethod
-    def __addToTree(cls, menu, item, checkFunction=None):
-        """ Helper function to recursively add items to a menu.
-        Add item (a dictionary that can contain more dictionaries) to menu
-        If check function is added will use it to check if the value must be added.
-        """
-        children = item.pop('children', [])
-
-        if checkFunction is not None:
-            add = checkFunction(item)
-            if not add:
-                return
-        subMenu = menu.addSubMenu(**item)  # we expect item={'text': ...}
-        for child in children:
-            cls.__addToTree(subMenu, child, checkFunction)  # add recursively to sub-menu
-
-        return subMenu
-
-    @classmethod
-    def __inSubMenu(cls, child, subMenu):
-        """
-        Return True if child belongs to subMenu
-        """
-        for ch in subMenu:
-            if child['tag'] == cls.TAG_PROTOCOL:
-                if not ch.value.empty() and ch.value == child['value']:
-                    return ch
-            elif ch.text == child['text']:
-                return ch
-        return None
-
-    @classmethod
-    def _orderSubMenu(cls, session):
-        """
-        Order all children of a given session:
-        The protocols first, then the sessions(the 'more' session at the end)
-        """
-        lengthSession = len(session.childs)
-        if lengthSession > 1:
-            childs = session.childs
-            lastChildPos = lengthSession - 1
-            if childs[lastChildPos].tag == cls.TAG_PROTOCOL:
-                for i in range(lastChildPos - 1, -1, -1):
-                    if childs[i].tag == cls.TAG_PROTOCOL:
-                        break
-                    else:
-                        tmp = childs[i + 1]
-                        childs[i + 1] = childs[i]
-                        childs[i] = tmp
-            else:
-                for i in range(lastChildPos - 1, -1, -1):
-                    if childs[i].tag == cls.TAG_PROTOCOL:
-                        break
-                    elif 'more' in str(childs[i].text).lower():
-                        tmp = childs[i + 1]
-                        childs[i + 1] = childs[i]
-                        childs[i] = tmp
-
-    @classmethod
-    def __findTreeLocation(cls, subMenu, children, parent):
-        """
-        Locate the protocol position in the given view
-        """
-        for child in children:
-            sm = cls.__inSubMenu(child, subMenu)
-            if sm is None:
-                cls.__addToTree(parent, child, cls.__checkItem)
-                cls._orderSubMenu(parent)
-            elif child['tag'] == cls.TAG_PROTOCOL_GROUP or child['tag'] == cls.TAG_SECTION:
-                cls.__findTreeLocation(sm.childs, child['children'], sm)
-
-    @classmethod
-    def __checkItem(cls, item):
-        """ Function to check if the protocol has to be added or not.
-        Params:
-            item: {"tag": "protocol", "value": "ProtImportMovies",
-                   "text": "import movies"}
-        """
-        if item["tag"] != cls.TAG_PROTOCOL:
-            return True
-
-        # It is a protocol as this point, get the class name and
-        # check if it is disabled
-        protClassName = item["value"]
-        protClass = pw.Config.getDomain().getProtocols().get(protClassName)
-
-        return False if protClass is None else not protClass.isDisabled()
-
-    @classmethod
-    def __addAllProtocols(cls, domain, protocols):
-        # Add all protocols
-        # FIXME: Check why this import is here
-        allProts = domain.getProtocols()
-
-        # Sort the dictionary
-        allProtsSorted = OrderedDict(sorted(allProts.items(),
-                                            key=lambda e: e[1].getClassLabel()))
-        allProtMenu = ProtocolConfig(cls.ALL_PROTOCOLS)
-        packages = {}
-
-        # Group protocols by package name
-        for k, v in allProtsSorted.items():
-            if cls.isAFinalProtocol(v, k):
-                packageName = v.getClassPackageName()
-                # Get the package submenu
-                packageMenu = packages.get(packageName)
-
-                # If no package menu available
-                if packageMenu is None:
-                    # Add it to the menu ...
-                    packageLine = {"tag": "package", "value": packageName,
-                                   "text": packageName}
-                    packageMenu = cls.__addToTree(allProtMenu, packageLine)
-
-                    # Store it in the dict
-                    packages[packageName] = packageMenu
-
-                # Add the protocol
-                tag = cls.getProtocolTag(v.isInstalled())
-
-                protLine = {"tag": tag, "value": k,
-                            "text": v.getClassLabel(prependPackageName=False)}
-
-                # If it's a new protocol
-                if v.isNew() and v.isInstalled():
-                    # add the new icon
-                    protLine["icon"] = "newProt.gif"
-
-                cls.__addToTree(packageMenu, protLine)
-
-        protocols[cls.ALL_PROTOCOLS] = allProtMenu
-
-    @classmethod
-    def __addProtocolsFromConf(cls, protocols, protocolsConfPath):
-        """
-        Load the protocols in the tree from a given protocols.conf file,
-        either the global one in Scipion or defined in a plugin.
-        """
-        # Populate the protocols menu from the plugin config file.
-        if os.path.exists(protocolsConfPath):
-            cp = ConfigParser()
-            cp.optionxform = str  # keep case
-            cp.read(protocolsConfPath)
-            #  Ensure that the protocols section exists
-            if cp.has_section('PROTOCOLS'):
-                for menuName in cp.options('PROTOCOLS'):
-                    if menuName not in protocols:  # The view has not been inserted
-                        menu = ProtocolConfig(menuName)
-                        children = json.loads(cp.get('PROTOCOLS', menuName))
-                        for child in children:
-                            cls.__addToTree(menu, child, cls.__checkItem)
-                        protocols[menuName] = menu
-                    else:  # The view has been inserted
-                        menu = protocols.get(menuName)
-                        children = json.loads(cp.get('PROTOCOLS',
-                                                     menuName))
-                        cls.__findTreeLocation(menu.childs, children, menu)
-
-    @classmethod
-    def load(cls, domain, protocolsConf):
-        """ Read the protocol configuration from a .conf file similar to the
-        one in scipion/config/protocols.conf,
-        which is the default one when no file is passed.
-        """
-        protocols = OrderedDict()
-        # Read the protocols.conf from Scipion (base) and create an initial
-        # tree view
-        cls.__addProtocolsFromConf(protocols, protocolsConf)
-
-        # Read the protocols.conf of any installed plugin
-        pluginDict = domain.getPlugins()
-        pluginList = pluginDict.keys()
-        for pluginName in pluginList:
-            try:
-
-                # if the plugin has a path
-                if isModuleAFolder(pluginName):
-                    # Locate the plugin protocols.conf file
-                    protocolsConfPath = os.path.join(
-                        pluginDict[pluginName].__path__[0],
-                        cls.PLUGIN_CONFIG_PROTOCOLS)
-                    cls.__addProtocolsFromConf(protocols, protocolsConfPath)
-
-            except Exception as e:
-                print('Failed to read settings. The reported error was:\n  %s\n'
-                      'To solve it, fix %s and run again.' % (
-                          e, os.path.abspath(protocolsConfPath)))
-
-            # Add all protocols to All view
-        cls.__addAllProtocols(pw.Config.getDomain(), protocols)
-
-        return protocols
